+++ lib/chewy/type/adapter/object.rb
@@ -17,18 +17,28 @@ module Chewy
           @type_name ||= (options[:name] || target).to_s.underscore
         end
 
+        # Imports passed data with options
+        #
+        # Import data types:
+        #
+        #   * Array ob objects
+        #
+        # Import options:
+        #
+        #   <tt>:batch_size</tt> - import batch size, 1000 objects by default
+        #
         def import *args, &block
           import_options = args.extract_options!
           batch_size = import_options.delete(:batch_size) || BATCH_SIZE
           objects = args.flatten
 
-          objects.in_groups_of(batch_size, false).all? do |group|
+          objects.in_groups_of(batch_size, false).map do |group|
             action_groups = group.group_by do |object|
               raise "Object is not a `#{target}`" if class_target? && !object.is_a?(target)
               object.respond_to?(:destroyed?) && object.destroyed? ? :delete : :index
             end
             block.call action_groups
-          end
+          end.all?
         end
 
         def load *args